equipmentRecords/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-phonenumber-field==7.0.0
djangorestframework==3.14.0
phonenumberslite==8.12.56
pytz==2022.2.1
sqlparse==0.4.2
tzdata==2022.2
